HTML,CSS, and Java-Script additions
Created a custom spotify holiday logo for us to use.
Cleaned up the font sizing for a cleaner look.
Added functionality to the "select a holiday" dropdown box so that itll show the corrisponding page to the right which will say the holiday name and will give us the correct format to add previews to songs and lyrics through spotify
Made the title hug the logo with a margin of 20px